Rename initTagField to TagController
